deletion

+++ domain/specialized_service/user_service/delegation_service.go
@@ -155,12 +155,21 @@ func (s *DelegationService) SpecializedDeleteRow(results []map[string]interface{
 		}
 		s.Domain.GetDb().DeleteQueryWithRestriction(ds.DBShare.Name, share, false)
 		res["state"] = "completed"
-		s.Domain.GetDb().DeleteQueryWithRestriction(ds.DBTask.Name, map[string]interface{}{
-			"binded_dbtask": s.Domain.GetDb().BuildSelectQueryWithRestriction(ds.DBTask.Name, map[string]interface{}{
-				ds.UserDBField: res[ds.UserDBField],
-			}, false, utils.SpecialIDParam),
-		}, false)
-		results[i] = task.SetClosureStatus(res)
+		arr := []interface{}{
+			connector.FormatSQLRestrictionWhereByMap("", map[string]interface{}{
+				ds.UserDBField: res["delegated_"+ds.UserDBField],
+			}, false),
+		}
+		currentTime := time.Now()
+		arr = append(arr, "('"+currentTime.Format("2006-01-02")+"' >= start_date AND ('"+currentTime.Format("2006-01-02")+"' < end_date OR end_date IS NULL))")
+		if rr, err := s.Domain.GetDb().SelectQueryWithRestriction(ds.DBDelegation.Name, arr, false); err == nil && len(rr) == 0 {
+			s.Domain.GetDb().DeleteQueryWithRestriction(ds.DBTask.Name, map[string]interface{}{
+				"binded_dbtask": s.Domain.GetDb().BuildSelectQueryWithRestriction(ds.DBTask.Name, map[string]interface{}{
+					ds.UserDBField: res["delegated_"+ds.UserDBField],
+				}, false, utils.SpecialIDParam),
+			}, false)
+			results[i] = task.SetClosureStatus(res)
+		}
 	}
 }
